Authentication

+++ std-dive-logger-backend/std-dive-logger-service/src/main/java/ch/sthomas/stddivelogger/service/UserService.java
@@ -13,7 +13,7 @@ public class UserService {
         this.userDataService = userDataService;
     }
 
-    public User getUserById(final int userId) {
+    public User getUserById(final long userId) {
         return userDataService.findUserById(userId);
     }
 }
